@@ -2,7 +2,7 @@ package com.alin.customapp.activity;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.TextView;
 import com.alin.customapp.R;
