@@ -1,6 +1,17 @@
 package com.alin.customapp.common;
 
-import android.support.v7.app.AppCompatActivity;
+import android.content.pm.ActivityInfo;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Toolbar;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import butterknife.ButterKnife;
+import cn.bingoogolapple.swipebacklayout.BGASwipeBackHelper;
+import com.alin.customapp.R;
+import com.alin.customapp.constant.AppStatusConstant;
+import com.alin.customapp.manager.AppStatusManager;
+import com.jaeger.library.StatusBarUtil;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -11,8 +22,142 @@ import java.io.InputStreamReader;
  * @author: Create By ZhangWenLin
  * @create: 2018-11-01 15:31
  **/
-public class BaseActivity extends AppCompatActivity {
+public abstract class BaseActivity extends AppCompatActivity implements BGASwipeBackHelper.Delegate, View.OnClickListener {
+    protected BGASwipeBackHelper mSwipeBackHelper;
+    protected Toolbar mToolbar;
 
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        // 设置闪屏页
+        setSplash();
+        // 「必须在 Application 的 onCreate 方法中执行 BGASwipeBackHelper.init 来初始化滑动返回」
+        // 在 super.onCreate(savedInstanceState) 之前调用该方法
+        initSwipeBackFinish();
+        super.onCreate(savedInstanceState);
+        ButterKnife.bind(this);
+        setScreenRoate(true);
+        setStatusBar();
+    }
+
+    /**
+     * 初始化滑动返回。在 super.onCreate(savedInstanceState) 之前调用该方法
+     */
+    private void initSwipeBackFinish() {
+        mSwipeBackHelper = new BGASwipeBackHelper(this, this);
+
+        // 「必须在 Application 的 onCreate 方法中执行 BGASwipeBackHelper.init 来初始化滑动返回」
+        // 下面几项可以不配置，这里只是为了讲述接口用法。
+
+        // 设置滑动返回是否可用。默认值为 true
+        mSwipeBackHelper.setSwipeBackEnable(true);
+        // 设置是否仅仅跟踪左侧边缘的滑动返回。默认值为 true
+        mSwipeBackHelper.setIsOnlyTrackingLeftEdge(true);
+        // 设置是否是微信滑动返回样式。默认值为 true
+        mSwipeBackHelper.setIsWeChatStyle(true);
+        // 设置阴影资源 id。默认值为 R.drawable.bga_sbl_shadow
+        mSwipeBackHelper.setShadowResId(R.drawable.bga_sbl_shadow);
+        // 设置是否显示滑动返回的阴影效果。默认值为 true
+        mSwipeBackHelper.setIsNeedShowShadow(true);
+        // 设置阴影区域的透明度是否根据滑动的距离渐变。默认值为 true
+        mSwipeBackHelper.setIsShadowAlphaGradient(true);
+        // 设置触发释放后自动滑动返回的阈值，默认值为 0.3f
+        mSwipeBackHelper.setSwipeBackThreshold(0.3f);
+        // 设置底部导航条是否悬浮在内容上，默认值为 false
+        mSwipeBackHelper.setIsNavigationBarOverlap(false);
+    }
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        ButterKnife.bind(this).unbind();
+    }
+
+    /**
+     * 是否支持滑动返回。这里在父类中默认返回 true 来支持滑动返回，如果某个界面不想支持滑动返回则重写该方法返回 false 即可
+     *
+     * @return
+     */
+    @Override
+    public boolean isSupportSwipeBack() {
+        return true;
+    }
+
+    /**
+     * 正在滑动返回
+     *
+     * @param slideOffset 从 0 到 1
+     */
+    @Override
+    public void onSwipeBackLayoutSlide(float slideOffset) {
+
+    }
+
+    /**
+     * 没达到滑动返回的阈值，取消滑动返回动作，回到默认状态
+     */
+    @Override
+    public void onSwipeBackLayoutCancel() {
+
+    }
+
+    /**
+     * 滑动返回执行完毕，销毁当前 Activity
+     */
+    @Override
+    public void onSwipeBackLayoutExecuted() {
+        mSwipeBackHelper.swipeBackward();
+    }
+
+    @Override
+    public void onBackPressed() {
+        // 正在滑动返回的时候取消返回按钮事件
+        if (mSwipeBackHelper.isSliding()) {
+            return;
+        }
+        mSwipeBackHelper.backward();
+    }
+
+    @Override
+    public void onClick(View view) {
+
+    }
+
+    /**
+     * 设置屏幕横竖屏切换
+     * @param screenRoate true  竖屏     false  横屏
+     */
+    private void setScreenRoate(Boolean screenRoate) {
+        if (screenRoate) {
+            //设置竖屏模式
+            setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
+        } else {
+            setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE);
+        }
+    }
+
+    /**
+     * 沉浸式实现
+     */
+    protected void setStatusBar() {
+        StatusBarUtil.setColorForSwipeBack(this, getResources().getColor(R.color.colorWhite, getTheme()), 0);
+        // StatusBarUtil.setTranslucentForImageViewInFragment(this, 0, null);
+    }
+
+    protected void setSplash() {
+        switch (AppStatusManager.getInstance().appStatus) {
+            case AppStatusConstant.STATUS_FORCE_KILLED:
+
+                break;
+            default:
+                break;
+        }
+    }
+
+    /**
+     * 获取assets文件字符串内容
+     * @param fileName 文件地址
+     * @return
+     */
     protected String getAssetsString(String fileName) {
         StringBuilder stringBuilder = new StringBuilder();
         try {
