@@ -5,9 +5,8 @@ import android.graphics.drawable.RotateDrawable;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 import android.widget.ImageView;
+import androidx.annotation.Nullable;
 import com.alin.customapp.R;
 import com.alin.customapp.common.BaseActivity;
 
