@@ -3,14 +3,13 @@ package com.alin.customapp.activity;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.KeyEvent;
 import android.view.View;
 import android.view.inputmethod.EditorInfo;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.*;
+import androidx.annotation.Nullable;
 import com.alibaba.fastjson.JSONObject;
 import com.alin.customapp.R;
 import com.alin.customapp.adapter.SearchActvAdapter;
