@@ -2,14 +2,10 @@ package com.alin.customapp.adapter;
 
 import android.content.Context;
 import android.os.Handler;
-import android.os.Message;
-import android.support.v4.view.PagerAdapter;
-import android.support.v4.view.ViewPager;
-import android.util.Log;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.ViewParent;
+import androidx.viewpager.widget.PagerAdapter;
 import com.alin.customapp.model.Banner;
 import com.alin.customapp.view.HttpImageView;
 
